+++ src/main/java/com/pjeskiem_i_mieczem/Warrior.java
@@ -8,15 +8,15 @@ public class Warrior extends Player{
 
     @Override
     public int getDamage() {
-        return this.strength.getValue() * 10 + luck.getValue();
+        return (int) (this.strength.getValue() * 10 + luck.getValue());
     }
 
     @Override
     public void takeDamage(Player other) {
-        float mage_reduction = (other instanceof Mage) ? intelligence.getValue() :0;
-        float warrior_reduction = (other instanceof Warrior) ? strength.getValue() :0;
-        float hunter_reduction = (other instanceof Hunter) ? dexterity.getValue() :0;
-        float shield_protection = strength.getValue();
+        float mage_reduction = (other instanceof Mage) ? (float) intelligence.getValue() :0;
+        float warrior_reduction = (other instanceof Warrior) ? (float) strength.getValue() :0;
+        float hunter_reduction = (other instanceof Hunter) ? (float) dexterity.getValue() :0;
+        float shield_protection = (float) strength.getValue();
 
     }
 }
